requirements.txt
click
coloredlogs
PyYaml>=5.1
